[FIX] project: avoid to remove the default `partner_id`
The onchange of the project field
removed the default partner loaded of the task
if there was no project set on the task, or if the
project was not associated to a specific partner.

There is no obvious reasons to do so,
and the revision 3fe06da63409d8015f739b6f601e976862370d33
introducing the regression doesn't explain why
it has been done. At first sight, there is no reason.

opw-657876