[FIX] web_kanban: properly removes nvd3 tooltips
with the recent update of nvd3 to the last version, we fixed a lot of
tooltip issues, but we introduced at least a new problem:  when a view
is detached from the dom with a visible tooltip, it no longer is cleaned
up.

In this commit, we simply remove all nvtooltips when we detach a kanban
view (necessary because dashboards inherits from kanban).  This is not
really an elegant way to solve the issue, but at least it works.  any
improvement is welcome.