[FIX] point_of_sale: don't ignore iface_tax_included
In 22c8db7fcb5bd85e1e78d52466bfcb126e499c2f the concept of
iface_tax_included was introduced. It was not actually used for the
display price though. It looks like a mistake during merging or
something because it adds the code responsible for dealing with the
option after a return statement in get_display_price().

opw-657983