[FIX] res.lang, tools: safeguards in case of inactive en_US
As of 5aa459a18e3c8a7759e7e6565ec5d97db7923251, en_US may not be
always active, e.g. when another language is used as main language.
Using en_US as fallback is not safe anymore, or at least we have
to check that it is active.

-> Since many parts of the code try to take the default/fallback
language from the "main company language", we need to make sure
to set it at installation time. (it doesn't use the defaults
as it is created in the bootstrap SQL)

-> In addition, when everything else fails, falling back to
the first active language is better than en_US. (e.g in formatLang)