[FIX] sale_timesheet: record employee cost if no SO line
When a user records an activity or a timesheet entry on an analytic
account not linked to a SO, the employee's cost is not taken into
account. This is due to the fact that the search of the timesheet cost
is only performed if there is a SO line.

This fix uncouples the SO line search with the timesheet cost search,
since there is no reason to link them.

opw-657899