[FIX] web: refresh in form view when first view mode is calendar
When you are in form view and came from calendar, a screen refresh
throwed an error from fullcalendar lib. Also when going back to the
calendar view, it would not be loaded.

It is due to fullcalendar not being rendered if it is not visible, and
other part of the Odoo code are dependent on it having been rendered.

This commit hide the calendar after having been rendered.

closes #10253
opw-659314