[FIX] website_quote:
This revision is related to 7be2403cf5ae54db485832fa33c0cf33645e0dc6

The payment acquirers must be fetched if the quote
needs a payment.

Besides, as explained in the above revision, the state
`manual` no longer exists sales orders, and therefore
this must no longer be used to determine if a quote
needs a payment or not.

opw-660575