[FIX] payment_buckaroo: return url, payment validation
This revision is related to 6efc371291951a3683b4aa30630781c478ea445f

The return URL parameter key has changed: The capital letters
are not the same than before.

Besides, send a dict in the `ADD_RETURNDATA` doesn't seem
to be supported by Buckaroo. We therefore no
longer uses a dict but a string to avoid problems.

opw-665697