[FIX] report: in report qweb, `doc` is an alias of `o`
In some reports,
e.g. the sale order report(report_saleorder),
the browse record is set under `doc` instead of
`o`.

Therefore, when printing a sale order as a user
from another company than the SO company,
the header printed were the one of the user company
instead of the one of the sale order company.

opw-666306